// Backend/Frontend LocalStorage/src/api/generated/ffbAPI.schemas.ts
export type Uuid = string;

export type AccountType = typeof AccountType[keyof typeof AccountType];


export const AccountType = {
  ADMIN: 'ADMIN',
  FOOD_COURT_WORKER: 'FOOD_COURT_WORKER',
  GUEST: 'GUEST',
} as const;
// ...
export interface LoginRequest {
  loginNr?: string;
  password?: string;
}
// ...
export interface RegisterRequest {
  loginNr?: string;
  password?: string;
}

export interface ShareOrderRequest {
  orderId?: Uuid;
  loginNr?: string;
}

export interface TicketRequest {
  loginNrs?: string[];
}

export interface TicketResponse {
  id?: Uuid;
  loginNr?: string;
}

export type GetAccountListAll400 = {
  code?: string;
  message?: string;
};

export type PostAccountLogin200 = {
  loginNr?: string;
  token?: string;
};

export type PostAccountLogin400 = {
  code?: string;
  message?: string;
};

export type PostAccountLogout400 = {
  code?: string;
  message?: string;
};

export type PostAccountRegister201 = {
  id?: Uuid;
  loginNr?: string;
  type?: AccountType;
};